making events clickable for admins only
client requested that only admins are able to click an event on the
calendar to view the full event details.

add an if statement to the bottom of the body which checks to see if
the admin is logged in, and then includes the appropriate js file to
initialize the calendar events.